+++ aws-sdk-core/features/cloudhsm/step_definitions.rb
@@ -1,0 +1,6 @@
+Before("@cloudhsm") do
+  @client = Aws::CloudHSM::Client.new
+end
+
+After("@cloudhsm") do
+end