+++ aws-sdk-core/lib/aws-sdk-core/cloudhsm.rb
@@ -1,0 +1,4 @@
+Aws.add_service(:CloudHSM, {
+  api: File.join(Aws::API_DIR, 'CloudHSM.api.json'),
+  docs: File.join(Aws::API_DIR, 'CloudHSM.docs.json'),
+})